Setup material and bavel and implement Tx of gcode
It is difficult to set up material and bavel depth of Path(s)
by the user due to too many Path(s).

This commit sets up a material of each Path(s) to easily set up
material by the user. The default material name is T0. And this
implement Tx of gcode (e.g. T0, T1 ... Tx) to change or to add
materials.

This commit also sets up default bavel depth is 0.2. This value
is good but not best, then which has room for discussion.

Signed-off-by: Mitsunori YOSHIDA marbocub@gmail.com