fix inconsistent line weight of a curve
The bavel of a curve often gets inconsistent line weight due to
lower U resolution to represent sharp corners.

This commit sets the length of handles to zero at control points
to represent sharp corners then sets upper the resolution to get
consistent line weight of the curve.

And which caps end face to close shape of the line.

Signed-off-by: Mitsunori YOSHIDA marbocub@gmail.com